Appease ruff and mypy
- Move resolve_canonical_ast/cst into a dedicated _canonical module so the
  base-class methods can import them at the top level (PLC0415).
- Flatten nested isinstance chain in get_matching_call_cst (SIM102).
- Reformat the import-tracker example table so ruff stops flagging the
  continuation line as commented-out code (ERA001).
- Inline the isinstance(ast.Call) check in critical_except so mypy's narrowing
  kicks in (attr-defined on "expr").
- Drop the now-unused identifier_to_string import from visitor91x.
